chore(flag): add new models for flags errors
