+++ src/features/discord/messages/classes/commands/flags/discord-command-boolean-flag.ts
@@ -1,5 +1,5 @@
 import { DiscordCommandFlagTypeEnum } from "../../../enums/commands/discord-command-flag-type.enum";
-import { IDiscordCommandFlag } from "../../../interfaces/commands/discord-command-flag";
+import { IDiscordCommandFlag } from "../../../interfaces/commands/flags/discord-command-flag";
 import { DiscordCommandFlag } from "./discord-command-flag";
 
 export class DiscordCommandBooleanFlag<
