+++ src/features/discord/messages/services/command/feature/services/feature-names/discord-message-command-feature-wrong-feature-name-error.service.spec.ts
@@ -6,20 +6,20 @@ import {
 } from "discord.js";
 import moment from "moment-timezone";
 import { createMock } from "ts-auto-mock";
-import { ColorEnum } from "../../../../../../../enums/color.enum";
-import { IconEnum } from "../../../../../../../enums/icon.enum";
-import { ServiceNameEnum } from "../../../../../../../enums/service-name.enum";
-import { CoreEventService } from "../../../../../../core/services/core-event.service";
-import { DiscordSoniaService } from "../../../../../users/services/discord-sonia.service";
-import { DiscordMessageCommandEnum } from "../../../../enums/commands/discord-message-command.enum";
-import { IAnyDiscordMessage } from "../../../../types/any-discord-message";
-import { DiscordMessageConfigService } from "../../../config/discord-message-config.service";
-import { DiscordMessageCommandCliErrorService } from "../../discord-message-command-cli-error.service";
-import { DISCORD_MESSAGE_COMMAND_FEATURE_NAMES } from "../constants/discord-message-command-feature-names";
-import { DiscordMessageCommandFeatureEmptyFeatureNameErrorService } from "./discord-message-command-feature-empty-feature-name-error.service";
+import { ColorEnum } from "../../../../../../../../enums/color.enum";
+import { IconEnum } from "../../../../../../../../enums/icon.enum";
+import { ServiceNameEnum } from "../../../../../../../../enums/service-name.enum";
+import { CoreEventService } from "../../../../../../../core/services/core-event.service";
+import { DiscordSoniaService } from "../../../../../../users/services/discord-sonia.service";
+import { DiscordMessageCommandEnum } from "../../../../../enums/commands/discord-message-command.enum";
+import { IAnyDiscordMessage } from "../../../../../types/any-discord-message";
+import { DiscordMessageConfigService } from "../../../../config/discord-message-config.service";
+import { DiscordMessageCommandCliErrorService } from "../../../discord-message-command-cli-error.service";
+import { DISCORD_MESSAGE_COMMAND_FEATURE_NAMES } from "../../constants/discord-message-command-feature-names";
+import { DiscordMessageCommandFeatureWrongFeatureNameErrorService } from "./discord-message-command-feature-wrong-feature-name-error.service";
 
-describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void => {
-  let service: DiscordMessageCommandFeatureEmptyFeatureNameErrorService;
+describe(`DiscordMessageCommandFeatureWrongFeatureNameErrorService`, (): void => {
+  let service: DiscordMessageCommandFeatureWrongFeatureNameErrorService;
   let coreEventService: CoreEventService;
   let discordSoniaService: DiscordSoniaService;
   let discordMessageConfigService: DiscordMessageConfigService;
@@ -33,20 +33,20 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
   });
 
   describe(`getInstance()`, (): void => {
-    it(`should create a DiscordMessageCommandFeatureEmptyFeatureNameError service`, (): void => {
+    it(`should create a DiscordMessageCommandFeatureWrongFeatureNameError service`, (): void => {
       expect.assertions(1);
 
-      service = DiscordMessageCommandFeatureEmptyFeatureNameErrorService.getInstance();
+      service = DiscordMessageCommandFeatureWrongFeatureNameErrorService.getInstance();
 
       expect(service).toStrictEqual(
-        expect.any(DiscordMessageCommandFeatureEmptyFeatureNameErrorService)
+        expect.any(DiscordMessageCommandFeatureWrongFeatureNameErrorService)
       );
     });
 
-    it(`should return the created DiscordMessageCommandFeatureEmptyFeatureNameError service`, (): void => {
+    it(`should return the created DiscordMessageCommandFeatureWrongFeatureNameError service`, (): void => {
       expect.assertions(1);
 
-      const result = DiscordMessageCommandFeatureEmptyFeatureNameErrorService.getInstance();
+      const result = DiscordMessageCommandFeatureWrongFeatureNameErrorService.getInstance();
 
       expect(result).toStrictEqual(service);
     });
@@ -61,14 +61,14 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
         .mockImplementation();
     });
 
-    it(`should notify the DiscordMessageCommandFeatureEmptyFeatureNameError service creation`, (): void => {
+    it(`should notify the DiscordMessageCommandFeatureWrongFeatureNameError service creation`, (): void => {
       expect.assertions(2);
 
-      service = new DiscordMessageCommandFeatureEmptyFeatureNameErrorService();
+      service = new DiscordMessageCommandFeatureWrongFeatureNameErrorService();
 
       expect(coreEventServiceNotifyServiceCreatedSpy).toHaveBeenCalledTimes(1);
       expect(coreEventServiceNotifyServiceCreatedSpy).toHaveBeenCalledWith(
-        ServiceNameEnum.DISCORD_MESSAGE_COMMAND_FEATURE_EMPTY_FEATURE_NAME_ERROR_SERVICE
+        ServiceNameEnum.DISCORD_MESSAGE_COMMAND_FEATURE_WRONG_FEATURE_NAME_ERROR_SERVICE
       );
     });
   });
@@ -76,6 +76,7 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
   describe(`getMessageResponse()`, (): void => {
     let anyDiscordMessage: IAnyDiscordMessage;
     let commands: DiscordMessageCommandEnum[];
+    let featureName: string;
 
     let discordMessageCommandCliErrorServiceGetCliErrorMessageResponseSpy: jest.SpyInstance;
     let discordSoniaServiceGetCorporationMessageEmbedAuthorSpy: jest.SpyInstance;
@@ -87,9 +88,10 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
     >;
 
     beforeEach((): void => {
-      service = new DiscordMessageCommandFeatureEmptyFeatureNameErrorService();
+      service = new DiscordMessageCommandFeatureWrongFeatureNameErrorService();
       anyDiscordMessage = createMock<IAnyDiscordMessage>();
       commands = [DiscordMessageCommandEnum.COOKIE];
+      featureName = `dummy-feature-name`;
 
       discordMessageCommandCliErrorServiceGetCliErrorMessageResponseSpy = jest.spyOn(
         discordMessageCommandCliErrorService,
@@ -126,7 +128,11 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
     it(`should get the CLI error message response`, async (): Promise<void> => {
       expect.assertions(2);
 
-      await service.getMessageResponse(anyDiscordMessage, commands);
+      await service.getMessageResponse(
+        anyDiscordMessage,
+        commands,
+        featureName
+      );
 
       expect(
         discordMessageCommandCliErrorServiceGetCliErrorMessageResponseSpy
@@ -149,7 +155,8 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
 
       const result = await service.getMessageResponse(
         anyDiscordMessage,
-        commands
+        commands,
+        featureName
       );
 
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -167,7 +174,8 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
 
       const result = await service.getMessageResponse(
         anyDiscordMessage,
-        commands
+        commands,
+        featureName
       );
 
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -182,7 +190,8 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
 
       const result = await service.getMessageResponse(
         anyDiscordMessage,
-        commands
+        commands,
+        featureName
       );
 
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -197,14 +206,15 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
 
       const result = await service.getMessageResponse(
         anyDiscordMessage,
-        commands
+        commands,
+        featureName
       );
 
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment
       // @ts-ignore
       expect(result.options.embed.fields[0]).toStrictEqual({
-        name: `Empty feature name`,
-        value: `You did not specify the name of the feature you wish to configure.\nI will not guess it for you so please try again with a feature name!\nAnd because I am kind and generous here is the list of all the features you can configure with an example.`,
+        name: `Wrong feature name`,
+        value: `\`dummy-feature-name\` is not an existing feature...\nLet me show you the list of available features with an example and maybe try again with a valid one this time, ok?`,
       } as EmbedFieldData);
     });
 
@@ -215,7 +225,8 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
 
       const result = await service.getMessageResponse(
         anyDiscordMessage,
-        commands
+        commands,
+        featureName
       );
 
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -241,7 +252,8 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
 
         const result = await service.getMessageResponse(
           anyDiscordMessage,
-          commands
+          commands,
+          featureName
         );
 
         // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -268,7 +280,8 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
 
         const result = await service.getMessageResponse(
           anyDiscordMessage,
-          commands
+          commands,
+          featureName
         );
 
         // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -300,7 +313,8 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
 
           const result = await service.getMessageResponse(
             anyDiscordMessage,
-            commands
+            commands,
+            featureName
           );
 
           // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -326,7 +340,8 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
 
           const result = await service.getMessageResponse(
             anyDiscordMessage,
-            commands
+            commands,
+            featureName
           );
 
           // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -359,7 +374,8 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
 
           const result = await service.getMessageResponse(
             anyDiscordMessage,
-            commands
+            commands,
+            featureName
           );
 
           // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -385,7 +401,8 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
 
           const result = await service.getMessageResponse(
             anyDiscordMessage,
-            commands
+            commands,
+            featureName
           );
 
           // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -406,7 +423,8 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
 
       const result = await service.getMessageResponse(
         anyDiscordMessage,
-        commands
+        commands,
+        featureName
       );
 
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -429,7 +447,8 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
 
         const result = await service.getMessageResponse(
           anyDiscordMessage,
-          commands
+          commands,
+          featureName
         );
 
         // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -453,7 +472,8 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
 
         const result = await service.getMessageResponse(
           anyDiscordMessage,
-          commands
+          commands,
+          featureName
         );
 
         // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -475,7 +495,8 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
 
       const result = await service.getMessageResponse(
         anyDiscordMessage,
-        commands
+        commands,
+        featureName
       );
 
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -492,7 +513,8 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
 
       const result = await service.getMessageResponse(
         anyDiscordMessage,
-        commands
+        commands,
+        featureName
       );
 
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -514,7 +536,8 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
 
       const result = await service.getMessageResponse(
         anyDiscordMessage,
-        commands
+        commands,
+        featureName
       );
 
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -531,7 +554,8 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
 
       const result = await service.getMessageResponse(
         anyDiscordMessage,
-        commands
+        commands,
+        featureName
       );
 
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -546,7 +570,8 @@ describe(`DiscordMessageCommandFeatureEmptyFeatureNameErrorService`, (): void =>
 
       const result = await service.getMessageResponse(
         anyDiscordMessage,
-        commands
+        commands,
+        featureName
       );
 
       expect(result.response).toStrictEqual(``);
