+++ src/features/discord/messages/types/commands/discord-command-flags-errors.ts
@@ -1,0 +1,3 @@
+import { IDiscordCommandFlagError } from "../../interfaces/commands/flags/discord-command-flag-error";
+
+export type IDiscordCommandFlagsErrors = IDiscordCommandFlagError[];
